+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/panels/MetricPanelBuilder.kt
@@ -3,6 +3,8 @@ package ru.yandex.money.tools.grafana.dsl.panels
 import org.json.JSONObject
 import ru.yandex.money.tools.grafana.dsl.datasource.Datasource
 import ru.yandex.money.tools.grafana.dsl.datasource.Graphite
+import ru.yandex.money.tools.grafana.dsl.generators.PanelLayoutGenerator
+import ru.yandex.money.tools.grafana.dsl.generators.SimplePanelLayoutGenerator
 import ru.yandex.money.tools.grafana.dsl.metrics.DashboardMetric
 import ru.yandex.money.tools.grafana.dsl.metrics.Metrics
 import ru.yandex.money.tools.grafana.dsl.metrics.MetricsBuilder
@@ -12,6 +14,8 @@ class MetricPanelBuilder(private val title: String) : PanelBuilder {
     private var propertiesSetter: (JSONObject) -> Unit = {}
 
     private val metrics = mutableListOf<DashboardMetric>()
+
+    private val panelLayoutGenerator: PanelLayoutGenerator = SimplePanelLayoutGenerator()
 
     override var bounds = PanelBuilder.DEFAULT_BOUNDS
 
@@ -29,7 +33,11 @@ class MetricPanelBuilder(private val title: String) : PanelBuilder {
 
     internal fun createPanel(): Panel = AdditionalPropertiesPanel(
             MetricPanel(
-                    BasePanel(id = idGenerator++, title = title, position = nextPosition(bounds.first, bounds.second)),
+                    BasePanel(
+                        id = panelLayoutGenerator.nextId(),
+                        title = title,
+                        position = panelLayoutGenerator.nextPosition(bounds.first, bounds.second)
+                    ),
                     datasource = datasource,
                     metrics = Metrics(metrics)
             ),