+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/panels/PanelsBuilder.kt
@@ -1,5 +1,6 @@
 package ru.yandex.money.tools.grafana.dsl.panels
 
+import ru.yandex.money.tools.grafana.dsl.generators.PanelLayoutGenerator
 import ru.yandex.money.tools.grafana.dsl.variables.Variable
 
 /**
@@ -8,7 +9,7 @@ import ru.yandex.money.tools.grafana.dsl.variables.Variable
  * @author Dmitry Komarov (komarovdmitry@yamoney.ru)
  * @since 7/21/18
  */
-class PanelsBuilder : PanelContainerBuilder {
+class PanelsBuilder(override val panelLayoutGenerator: PanelLayoutGenerator) : PanelContainerBuilder {
 
     override val panels = mutableListOf<Panel>()
 
@@ -20,7 +21,7 @@ class PanelsBuilder : PanelContainerBuilder {
     }
 
     fun row(title: String, repeatFor: Variable? = null, build: RowBuilder.() -> Unit) {
-        val builder = RowBuilder(title, repeatFor?.name)
+        val builder = RowBuilder(title, repeatFor?.name, panelLayoutGenerator)
         builder.build()
         panels += builder.createRow()
     }