+++ src/test/kotlin/ru/yandex/money/tools/grafana/dsl/panels/RowBuilderTest.kt
@@ -4,6 +4,7 @@ import org.amshove.kluent.shouldBe
 import org.testng.annotations.Test
 import ru.yandex.money.tools.grafana.dsl.dashboard.DashboardBuilder
 import ru.yandex.money.tools.grafana.dsl.datasource.Zabbix
+import ru.yandex.money.tools.grafana.dsl.generators.SimplePanelLayoutGenerator
 import ru.yandex.money.tools.grafana.dsl.jsonFile
 import ru.yandex.money.tools.grafana.dsl.shouldEqualToJson
 
@@ -12,7 +13,7 @@ class RowBuilderTest : AbstractPanelTest() {
     @Test
     fun `should create row with 1 panel`() {
         // given
-        val panelsBuilder = PanelsBuilder()
+        val panelsBuilder = PanelsBuilder(SimplePanelLayoutGenerator())
         val dashboardBuilder = DashboardBuilder("Test Dashboard")
         val values by dashboardBuilder.variables.query(datasource = Zabbix, query = "My variable") {
             regex = ".*"