+++ src/test/kotlin/ru/yandex/money/tools/grafana/dsl/panels/TestContainerBuilder.kt
@@ -1,5 +1,8 @@
 package ru.yandex.money.tools.grafana.dsl.panels
 
-class TestContainerBuilder : PanelContainerBuilder {
+import ru.yandex.money.tools.grafana.dsl.generators.PanelLayoutGenerator
+import ru.yandex.money.tools.grafana.dsl.generators.SimplePanelLayoutGenerator
+
+class TestContainerBuilder(override val panelLayoutGenerator: PanelLayoutGenerator = SimplePanelLayoutGenerator()) : PanelContainerBuilder {
     override val panels = mutableListOf<Panel>()
 }